+++ src/resources/mod.rs
@@ -13,7 +13,7 @@ use controls::AnyHandle;
 use error::Error;
 
 pub use self::font::{FontT, Font};
-pub use self::image::{ImageT, Image};
+pub use self::image::{ImageT, OemImageT, Image};
 
 /**
     Structures implementing this trait can be used by a Ui to build a Resource